Fix styled-components test for Turbopack
